Move example objects out of DelaunayUtil
